fix backend for delete function

+++ backend/src/main/java/dev/wardrobe/WardrobeCollectionManager/DeleteItemService.java
@@ -2,9 +2,7 @@ package dev.wardrobe.WardrobeCollectionManager;
 
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.stereotype.Service;
-import dev.wardrobe.WardrobeCollectionManager.AddItemDTO;
 
 @Service
 public class DeleteItemService {
@@ -16,17 +14,24 @@ public class DeleteItemService {
         this.itemRepository = itemRepository;
     }
 
+    // Delete by ObjectId
     public void deleteById(ObjectId _id) {
-        itemRepository.deleteById(_id);
+        if (itemRepository.existsById(_id)) {
+            itemRepository.deleteById(_id);
+        } else {
+            throw new RuntimeException("Item not found with ID: " + _id.toString());
+        }
     }
 
+    // Delete all items
     public void deleteAll() {
         itemRepository.deleteAll();
     }
 
-    public void deleteByName(String itemId) {
+    // This method is not needed now if you're using ObjectId directly in the controller
+    public void deleteByName(String _id) {
         itemRepository.findAll().stream()
-                .filter(item -> item.getItemId().equals(itemId))
+                .filter(item -> item.get_id().equals(new ObjectId(_id)))
                 .forEach(item -> itemRepository.deleteById(item.get_id()));
     }
 }
